refactor: StudyUtil 클래스에 맞지 않는 메서드 이동
Response로 변환하는 메서드를 Mapper 클래스로 이동
